Update email.py
Fix for bug causing exception when email to be sent has URL attachments. Current code treats URL attachments as a local file. The style I added to the link tricks GMail to show a nice large icon preview of the document at the link attached. If not using Gmail it will show a box around the link making it look like a file attachment.